Table API Refactor
-Make Table class abstract, move existing implementation to SystemTable class.
