+++ src/controller/CMakeLists.txt
@@ -22,6 +22,7 @@ list(APPEND rocprofvis_controller_source_files
     src/rocprofvis_controller_trace.cpp
     src/rocprofvis_controller_track.cpp
     src/rocprofvis_controller_table.cpp
+	src/rocprofvis_controller_table_system.cpp
     src/rocprofvis_controller.cpp
     src/rocprofvis_controller_flow_control.cpp
     src/rocprofvis_controller_call_stack.cpp
