+++ src/controller/src/rocprofvis_controller_table.h
@@ -8,7 +8,6 @@
 #include "rocprofvis_c_interface.h"
 #include <vector>
 #include <map>
-#include <deque>
 #include <string>
 
 namespace RocProfVis
@@ -18,7 +17,6 @@ namespace Controller
 
 class Arguments;
 class Array;
-class Track;
 
 class Table : public Handle
 {
@@ -27,38 +25,21 @@ public:
 
     virtual ~Table();
 
-    rocprofvis_controller_object_type_t GetType(void) final;
+    virtual rocprofvis_controller_object_type_t GetType(void);
 
-    void Reset();
+    virtual rocprofvis_result_t Setup(rocprofvis_dm_trace_t dm_handle, Arguments& args) = 0;
+    virtual rocprofvis_result_t Fetch(rocprofvis_dm_trace_t dm_handle, uint64_t index,
+                              uint64_t count, Array& array) = 0;
 
-    rocprofvis_result_t Setup(rocprofvis_dm_trace_t dm_handle, Arguments& args);
-    rocprofvis_result_t Fetch(rocprofvis_dm_trace_t dm_handle, uint64_t index,
-                              uint64_t count, Array& array);
-
-    // Handlers for getters.
-    rocprofvis_result_t GetUInt64(rocprofvis_property_t property, uint64_t index, uint64_t* value) final;
-    rocprofvis_result_t GetDouble(rocprofvis_property_t property, uint64_t index, double* value) final;
-    rocprofvis_result_t GetObject(rocprofvis_property_t property, uint64_t index, rocprofvis_handle_t** value) final;
-    rocprofvis_result_t GetString(rocprofvis_property_t property, uint64_t index, char* value, uint32_t* length) final;
-
-    rocprofvis_result_t SetUInt64(rocprofvis_property_t property, uint64_t index, uint64_t value) final;
-    rocprofvis_result_t SetDouble(rocprofvis_property_t property, uint64_t index, double value) final;
-    rocprofvis_result_t SetObject(rocprofvis_property_t property, uint64_t index, rocprofvis_handle_t* value) final;
-    rocprofvis_result_t SetString(rocprofvis_property_t property, uint64_t index, char const* value, uint32_t length) final;
-
-private:
+protected:
     struct ColumnDefintion
     {
         std::string m_name;
         rocprofvis_controller_primitive_type_t m_type;
     };
 
-    std::vector<uint32_t> m_tracks;
     std::vector<ColumnDefintion> m_columns;
     std::map<uint64_t, std::vector<Data>> m_rows;
-    std::deque<uint64_t> m_lru;
-    double m_start_ts;
-    double m_end_ts;
     uint64_t m_num_items;
     uint64_t m_id;
     uint64_t m_sort_column;
